requirements.txt
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
decorator==4.0.10
diff-match-patch==20121119
Django==1.9.9
django-appconf==1.0.2
django-autocomplete-light==3.1.8
django-ckeditor==5.1.1
django-compressor==2.0
django-debug-toolbar==1.5
django-extra-views==0.6.4
django-filer==1.2.5
django-haystack==2.4.1
django-htmlmin==0.9.1
django-import-export==0.5.0
django-mptt==0.8.6
django-oscar==1.2.1
django-polymorphic==0.8.1
django-tables2==1.0.7
django-treebeard==4.0.1
django-widget-tweaks==1.4.1
easy-thumbnails==2.3
factory-boy==2.6.1
fake-factory==0.5.7
FeinCMS==1.12.1
funcsigs==1.0.2
html5lib==0.999999999
ipaddress==1.0.16
ipython==5.1.0
ipython-genutils==0.1.0
mock==1.3.0
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
phonenumbers==7.4.3
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
psycopg2==2.6.1
ptyprocess==0.5.1
purl==1.3
pycountry==1.20
Pygments==2.1.3
pysolr==3.5.0
python-dateutil==2.5.3
python-memcached==1.58
pytz==2016.4
rcssmin==1.0.6
requests==2.10.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.2.1
tablib==0.11.2
traitlets==4.2.2
Unidecode==0.4.19
wcwidth==0.1.7
webencodings==0.5
wheel==0.24.0
Whoosh==2.7.4
